// src/api/generated/models/getCompaniesCompanyIdRequestsInboundParams.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { GetCompaniesCompanyIdRequestsInboundStatus } from './getCompaniesCompanyIdRequestsInboundStatus';
import type { GetCompaniesCompanyIdRequestsInboundPriority } from './getCompaniesCompanyIdRequestsInboundPriority';
import type { GetCompaniesCompanyIdRequestsInboundSortBy } from './getCompaniesCompanyIdRequestsInboundSortBy';
import type { GetCompaniesCompanyIdRequestsInboundSortOrder } from './getCompaniesCompanyIdRequestsInboundSortOrder';

export type GetCompaniesCompanyIdRequestsInboundParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
/**
 * @minimum 0
 * @maximum 9007199254740991
 */
offset?: number;
status?: GetCompaniesCompanyIdRequestsInboundStatus;
priority?: GetCompaniesCompanyIdRequestsInboundPriority;
/**
 * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
 */
buyerCompanyId?: string;
/**
 * @minLength 1
 */
q?: string;
distributedFrom?: unknown;
distributedTo?: unknown;
sortBy?: GetCompaniesCompanyIdRequestsInboundSortBy;
sortOrder?: GetCompaniesCompanyIdRequestsInboundSortOrder;
};