// src/api/generated/models/getCompaniesCompanyIdRequestsInboundPriority.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type GetCompaniesCompanyIdRequestsInboundPriority = typeof GetCompaniesCompanyIdRequestsInboundPriority[keyof typeof GetCompaniesCompanyIdRequestsInboundPriority];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetCompaniesCompanyIdRequestsInboundPriority = {
  LOW: 'LOW',
  NORMAL: 'NORMAL',
  HIGH: 'HIGH',
  URGENT: 'URGENT',
} as const;