// src/api/generated/models/getCompaniesCompanyIdRequestsParams.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */
import type { GetCompaniesCompanyIdRequestsStatus } from './getCompaniesCompanyIdRequestsStatus';
import type { GetCompaniesCompanyIdRequestsPriority } from './getCompaniesCompanyIdRequestsPriority';
import type { GetCompaniesCompanyIdRequestsSortBy } from './getCompaniesCompanyIdRequestsSortBy';
import type { GetCompaniesCompanyIdRequestsSortOrder } from './getCompaniesCompanyIdRequestsSortOrder';

export type GetCompaniesCompanyIdRequestsParams = {
/**
 * @minimum 1
 * @maximum 100
 */
limit?: number;
/**
 * @minimum 0
 * @maximum 9007199254740991
 */
offset?: number;
status?: GetCompaniesCompanyIdRequestsStatus;
/**
 * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
 */
productId?: string;
/**
 * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
 */
assigneeUserId?: string;
/**
 * @pattern ^([0-9a-fA-F]{8}-[0-9a-fA-F]{4}-[1-8][0-9a-fA-F]{3}-[89abAB][0-9a-fA-F]{3}-[0-9a-fA-F]{12}|00000000-0000-0000-0000-000000000000|ffffffff-ffff-ffff-ffff-ffffffffffff)$
 */
createdByUserId?: string;
priority?: GetCompaniesCompanyIdRequestsPriority;
/**
 * @minLength 1
 */
q?: string;
createdFrom?: unknown;
createdTo?: unknown;
sortBy?: GetCompaniesCompanyIdRequestsSortBy;
sortOrder?: GetCompaniesCompanyIdRequestsSortOrder;
};