// src/api/generated/models/getCompaniesCompanyIdRequestsPriority.ts
/**
 * /** Generated by Orval — do not edit. Run: npm run codegen *\/
 */

export type GetCompaniesCompanyIdRequestsPriority = typeof GetCompaniesCompanyIdRequestsPriority[keyof typeof GetCompaniesCompanyIdRequestsPriority];


// eslint-disable-next-line @typescript-eslint/no-redeclare
export const GetCompaniesCompanyIdRequestsPriority = {
  LOW: 'LOW',
  NORMAL: 'NORMAL',
  HIGH: 'HIGH',
  URGENT: 'URGENT',
} as const;